requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
Authlib==1.3.1
blinker==1.9.0
blis==0.7.11
cachelib==0.13.0
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.7
cloudpathlib==0.21.0
colorama==0.4.6
confection==0.1.5
contourpy==1.3.2
cryptography==44.0.2
cssbeautifier==1.15.4
cycler==0.12.1
cymem==2.0.11
dataclasses-json==0.6.7
deprecation==2.1.0
djlint==1.36.4
docker==7.1.0
EditorConfig==0.17.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
Faker==37.1.0
filelock==3.18.0
Flask==3.1.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
fonttools==4.58.0
frozenlist==1.5.0
fsspec==2025.3.2
fuzzywuzzy==0.18.0
greenlet==3.1.1
grpcio==1.71.0
grpcio-health-checking==1.71.0
grpcio-tools==1.71.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.30.2
idna==3.10
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
jsbeautifier==1.15.4
json5==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
keybert==0.9.0
kiwisolver==1.4.8
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.51
langchain-ollama==0.3.1
langchain-text-splitters==0.3.8
langchain-weaviate==0.0.4
langcodes==3.5.0
langsmith==0.3.28
language_data==1.3.0
Levenshtein==0.27.1
lxml==5.4.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
mdurl==0.1.2
mpmath==1.3.0
msgspec==0.19.0
multidict==6.2.0
murmurhash==1.0.12
mypy-extensions==1.0.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
ollama==0.4.7
orjson==3.10.16
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pdfminer.six==20250327
pdfplumber==0.11.6
pefile==2023.2.7
pillow==11.1.0
preshed==3.0.9
prettytable==3.16.0
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
pyflakes==3.3.2
Pygments==2.19.1
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.3
pyparsing==3.2.3
PyPDF2==3.0.1
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-Levenshtein==0.27.1
pytz==2025.2
pywin32==310
pywin32-ctypes==0.2.3
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
semver==3.0.4
sentence-transformers==4.0.2
setuptools==78.1.0
shellingham==1.5.4
simsimd==6.2.1
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.40
srsly==2.5.1
sympy==1.13.1
tabulate==0.9.0
tenacity==9.1.2
thinc==8.2.5
threadpoolctl==3.6.0
tk==0.1.0
tokenizers==0.21.1
torch==2.6.0+cu126
torchaudio==2.6.0+cu126
torchvision==0.21.0+cu126
tqdm==4.67.1
transformers==4.51.1
typer==0.15.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
validators==0.34.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
weaviate-cli==3.2.0
weaviate-client==4.13.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
yamllint==1.37.1
yarl==1.19.0
zstandard==0.23.0
